requirements.txt
nbconvert
nbformat
pyyaml
notebook
requests
python-dotenv
pandas
seaborn
scikit-learn
progress
mammoth
pillow
python-docx
markdownify